Fix build issue introduced in d1e1eae
dpkg-shlibdeps couldn't find appimagelauncherfs's dependency on
libappimage, due to the RPATH not being set. This wasn't necessary until
libappimage was introduced as a dependency.

Unfortunately, dpkg-shlibdeps even fails on this if you run it on any
other binary in the install path. Took me a while to figure out the real
issue.

Thanks to git bisect! Could easily automate the check (run cmake, make
preinstall, then cpack). This really helped track the issue down,
especially finding out that the breaking change was about the filesystem
and it being linked to a new library. ldd-ing the binary showed the
broken/missing dependencies.