Extend and improve settings, mk. 5
Make sure the settings open with the first tab. Sometimes when you work
on a dialog in Qt Designer, you are on some tab other than the first
one. Qt will then launch with that tab selected. Since I don't want to
take care of saving with the right tab selected, I wanted to make sure
code wise that the right tab is shown on startup.