Added zh_CN language code variant for Simplified Chinese translation
Added zh_CN language code variant for Simplified Chinese translation. The current zh_Hans language code is technically "not-wrong", but practically never used by every system I tested on. As a result, Appimagelauncher does not show any translation for Chinese locale, be it Simplified or Traditional. The most common language code for Simplified Chinese is zh_CN. This commit is aimed to solve this issue.