Update BUILD instructions
Maintainer of primary repo has said to use "Master" instead of the Stable branch because of dltray, etc. problems.  Further, `sudo make install` *can* in fact harm your system, but if you are on any mainstream distribution, `checkinstall` is actually available in most repositories and builds a native-package-manager-compatibile package around what `sudo make install` does and then *installs* that package so you can easily revert afterwards.  This tends to be the safer option.