Extract binfmt bypass algorithm into library
This new library will be used to implement a new binfmt interpreter, which has the potential to finally resolve the mess that is the combination of AppImageLauncher and Docker.

Extracting the code into a library also gives us a perspective on eliminating the need for the binfmt-bypass tool as a standalone library, since the functionality could be integrated directly into AppImageLauncher.

The commit further fixes a variety of minor bugs and issues.