Add new binfmt interpreter binary
This interpreter introduces a new layer between binfmt_misc and AppImageLauncher. The binary dynamically checks whether AppImageLauncher is available (within the current cgroup/namespace), and if it is not, launches the AppImage using the same algorithm binfmt-bypass uses. This should effectively fix all issues related to using AppImages in Docker containers with AppImageLauncher installed on the host system, since we use the F (fix binary) mode with binfmt, which makes sure the kernel loads the binary from the host once instead of looking for it upon every launch in every cgroup/namespace/chroot where it may not be available.
In the long term, we could even optimize the launch times of AppImages by, e.g., implementing APPIMAGELAUNCHER_DISABLE in this tool as well.

This commit also eliminates the update-binfmts related workaround, as systemd-binfmt.service appears to work reliably now.

We still provide binfmt-bypass because some people may use it as a workaround currently, and it is useful for testing.