Embed preload libs in bypass tool
This commit allows us to use the tool within environments that do not contain the preload libs (e.g., changeroots, Docker containers). By compiling in the contents, we can create temporary preload library files which we can use to launch the AppImage.

This first implementation is not really efficient (it creates a file for every execution), but may be optimized eventually.