Rework updater
The new updater still uses the dialog from libappimageupdate-qt (which needs some updates to be more usable overall). However, the update check is now integrated in a more "wizard style" user interface that is shown right away with a busy indicator. Update check errors are also handled in that same window.

Overall, these features should eventually end up in libappimageupdate-qt, which should generally be rewritten to be non-blocking in all states.